Cleanup

+++ disk_usage.py
@@ -1,6 +1,8 @@
 import os
 import datetime
+import shutil
 import syslog
+import traceback
 
 import os_help
 import surveillance_config as config
@@ -18,8 +20,8 @@ def get_dir_size(path):
 class UsageInfo:
     def __init__(self, directory):
         self.directory = directory
-        self.date = datetime.date.strptime(directory, 
-                                           config.directory_date_format)
+        self.date = datetime.datetime.strptime(
+            directory, config.directory_date_format)
         
         all_path = os.path.join(config.video_store_path, directory, 
                                 config.video_all_dir)
@@ -28,17 +30,21 @@ class UsageInfo:
             self.bytes_used_all = get_dir_size(all_path)
         
         motion_path = os.path.join(config.video_store_path, directory,
-                                  config.video_all_dir)
-        self.bytes_used_motion = get_dir_size(video_all_dir)
+                                  config.video_motion_dir)
+        self.bytes_used_motion = get_dir_size(motion_path)
         
 class UsageInfos:
     """ Stores usage info for each directory, newest directories first, oldest last."""
     def __init__(self):
+        
         self.usage_info_list = []
         for item in os.listdir(config.video_store_path):
-            if not os.path.isdir(item): continue
+            if (item == config.video_unprocessed_dir 
+                or not os.path.isdir(
+                   os.path.join(config.video_store_path, item))):
+                continue
             try:
-                usage_info_list.append(UsageInfo(item));
+                self.usage_info_list.append(UsageInfo(item))
             except:
                 syslog.syslog(1, traceback.format_exc())
                 traceback.print_exc()
@@ -70,8 +76,9 @@ class UsageInfos:
         
 def clear_non_motion_segments_for_days_too_old(usage_infos):
     now = datetime.datetime.now()
-    for info in usage_infos.usage_info_list:
-        if (now - info.date) > config.duration_to_keep_all_segments:
+    for info in reversed(usage_infos.usage_info_list):
+        if (info.bytes_used_all != None and
+            (now - info.date) > config.duration_to_keep_all_segments):
             syslog.syslog(2, "Clearing non-motion to free disk space: " 
                   + info.directory)
             shutil.rmtree(os.path.join(config.video_store_path, 
@@ -88,15 +95,21 @@ def remove_oldest_day(usage_infos):
         config.video_store_path, usage_infos.get_oldest_day().directory), True);
     usage_infos.remove_oldest_day()
 
-
 def cleanup():
     """ Keeps all segments as specified, and cleans up others to keep total usage in check. """
-    usage_infos = UsageInfos()
-    clear_non_motion_segments_for_days_too_old(usage_infos)
+    infos = UsageInfos()
     
-    max_day_size = usage_infos.get_max_all_usage()
-    while (usage_infos.compute_space_usaged_ignoring_today() + max_day_size >
-           config.disk_usage_limit_bytes):
-        remove_oldest_day(usage_infos)
+    on_disk = infos.compute_space_usaged_ignoring_today()
+    max_day_size = infos.get_max_all_usage()
+    syslog.syslog(3, "Disk Usage initial (except today): " + str(on_disk)
+                  + " Max Day Size: " + str(max_day_size))
+    
+    clear_non_motion_segments_for_days_too_old(infos)
+    on_disk = infos.compute_space_usaged_ignoring_today()
+    while (on_disk + max_day_size > config.disk_usage_limit_bytes):
+        remove_oldest_day(infos)
+        on_disk = infos.compute_space_usaged_ignoring_today()
+        
+    syslog.syslog(3, "Disk Usage final (except today): " + str(on_disk))
     
     return
